pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "fastapi-backend"
readme = "README.md"
requires-python = ">=3.12"

# 生产运行时依赖：只保留真正业务需要 import 的模块
dependencies = [
    "fastapi",
    "uvicorn",
    "starlette>0.49.1",
    "pydantic",
    "annotated-types",
    "anyio>=4.10.0",
    "httpx",
    "h11",
    "httpcore",
    "idna",
    "SQLAlchemy",
    "aiomysql>0.3.0",
    "PyMySQL",
    "numpy",
    "pandas",
    "celery",
    "APScheduler",
    "amqp",
    "kombu",
    "billiard",
    "vine",
    "celery-types",
    "redis",
    "loguru",
    "PyJWT",
    "bcrypt",
    "python-dotenv",
    "shortuuid",
    "snowflake-id",
    "openai",
    "orjson",
    "tqdm",
    "jiter",
    "protobuf",
    "grpcio",
    "grpcio-tools",
    "certifi",
    "click",
    "packaging",
    "python-dateutil",
    "pytz",
    "tzdata",
    "tzlocal",
    "typing-extensions",
    "typing-inspection",
    "xxhash",
    "uvloop>=0.22.1; sys_platform != 'win32'",
    "httptools",
    "async-lru>=2.0.5",
    "celery-redbeat>=2.3.3",
    "trio>=0.32.0",
    "pydantic-settings>=2.12.0",
    "cryptography>=46.0.3",
    "ruff>=0.14.6",
    "oss2>=2.19.1",
    "boto3>=1.42.6",
    "boto3-stubs>=1.42.6",
    "uuid6>=2025.0.1",
    "pyyaml>=6.0.3",
    "aiosqlite>=0.22.1",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "opentelemetry-instrumentation-redis>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
]

# 分组依赖 —— 开发环境才需要
[dependency-groups]
dev = [
    # 构建工具也放入 dev，方便本地 build（但生产以 build-system 为准）
    "pip>=25.3",
    "setuptools>=79.0.1",
    "wheel>=0.45.1",
    # 常规开发工具（可根据你需要继续添加）
    "pytest",
    "pytest-asyncio",
    "httpx", # 测试请求常用
    "black",
    "ruff",
    "mypy",
    "types-redis",
    "types-pytz",
    "pandas-stubs>=2.3.3.251201",
    "twisted>=25.5.0",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
    "basedpyright>=1.37.4",
]

# CI 构建相关工具（可选，按需添加）
build = [
    "build", # python -m build
    "twine", # 上传包（如果你发包）
]

[build-system]
requires = [
    "setuptools>=79.0.1",
    "wheel>=0.45.1",
]
build-backend = "setuptools.build_meta"

[tool.uv]
package = false
default-groups = []

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
python-downloads = false

[tool.ruff]
# 核心规则
line-length = 120
target-version = "py312" # 假设你用 Python 3.12+

[tool.ruff.lint]
# 启用规则组合：
# E/W: Pycodestyle
# F: Pyflakes
# I: Isort (导入排序)
# B: Bugbear (发现潜在 bug)
# UP: Pyupgrade (自动升级语法)
# FAST: FastAPI 特有规则 (如果 Ruff 版本够新)
select = ["E", "W", "F", "I", "B", "UP"]

# 忽略规则：
# E501: 行太长由 formatter 处理，不需要 linter 报错
ignore = ["E501"]

[tool.ruff.format]
# 格式化风格
quote-style = "double"
indent-style = "space"
# 保持魔法尾随逗号功能 (True=禁用, False=启用)
# 也就是你想要的: 加了逗号就换行，没加就折叠
skip-magic-trailing-comma = false
# 行尾符
line-ending = "auto"

[tool.ruff.lint.isort]
# 导入分组设置
combine-as-imports = true

[tool.pytest.ini_options]
markers = [
    "integration: 集成测试，需要完整的 Celery Worker 和 Redis 环境",
]
